+++ include/md/forcefield/plane_surface_forcefield.hpp
@@ -1,4 +1,4 @@
-// Copyright snsinfu 2019.
+// Copyright snsinfu 2019-2021.
 // Distributed under the Boost Software License, Version 1.0. (See accompanying
 // file LICENSE.txt or copy at http://www.boost.org/LICENSE_1_0.txt)
 
@@ -176,9 +176,9 @@ namespace md
         }
 
         inline
-        auto plane_inward_potential(md::system const&, md::index i) const
+        auto plane_inward_potential(md::system const& system, md::index i) const
         {
-            return potfun_(i);
+            return potfun_(system, i);
         }
 
     private:
@@ -197,9 +197,9 @@ namespace md
         }
 
         inline
-        auto plane_outward_potential(md::system const&, md::index i) const
+        auto plane_outward_potential(md::system const& system, md::index i) const
         {
-            return potfun_(i);
+            return potfun_(system, i);
         }
 
     private:
@@ -212,7 +212,7 @@ namespace md
     template<typename P>
     auto make_plane_inward_forcefield(P pot)
     {
-        auto potfun = detail::make_field_potfun(pot);
+        auto potfun = detail::make_field_potential_factory(pot);
         using potfun_type = decltype(potfun);
         return md::basic_plane_inward_forcefield_impl<potfun_type>{potfun};
     }
@@ -223,7 +223,7 @@ namespace md
     template<typename P>
     auto make_plane_outward_forcefield(P pot)
     {
-        auto potfun = detail::make_field_potfun(pot);
+        auto potfun = detail::make_field_potential_factory(pot);
         using potfun_type = decltype(potfun);
         return md::basic_plane_outward_forcefield_impl<potfun_type>{potfun};
     }