+++ include/md/forcefield/point_source_forcefield.hpp
@@ -1,4 +1,4 @@
-// Copyright snsinfu 2018.
+// Copyright snsinfu 2018-2021.
 // Distributed under the Boost Software License, Version 1.0. (See accompanying
 // file LICENSE.txt or copy at http://www.boost.org/LICENSE_1_0.txt)
 
@@ -127,9 +127,9 @@ namespace md
         {
         }
 
-        auto point_source_potential(md::system const&, md::index i) const
+        auto point_source_potential(md::system const& system, md::index i) const
         {
-            return potfun_(i);
+            return potfun_(system, i);
         }
 
     private:
@@ -141,7 +141,7 @@ namespace md
     template<typename P>
     auto make_point_source_forcefield(P pot)
     {
-        auto potfun = detail::make_field_potfun(pot);
+        auto potfun = detail::make_field_potential_factory(pot);
         using potfun_type = decltype(potfun);
         return md::basic_point_source_forcefield<potfun_type>{potfun};
     }