+++ tests/forcefield/test_plane_surface_forcefield.cc
@@ -187,47 +187,48 @@ TEST_CASE("make_plane_inward_forcefield - creates a plane_surface_forcefield")
 {
     SECTION("fixed potential")
     {
-        md::harmonic_potential potential{1.23};
-        auto ff =
-            md::make_plane_inward_forcefield(potential)
-            .set_plane(
-                md::plane{md::vector{0, 0, 1}, md::point{0, 0, 0}}
-            );
+        auto ff = md::make_plane_inward_forcefield(
+            md::harmonic_potential{42}
+        );
 
         md::system system;
         md::harmonic_potential pot = ff.plane_inward_potential(system, 0);
-        CHECK(pot.spring_constant == Approx(potential.spring_constant));
+        CHECK(pot.spring_constant == 42);
 
         using ff_type = decltype(ff);
         CHECK(std::is_base_of<md::plane_surface_forcefield<ff_type>, ff_type>::value);
     }
 
-    SECTION("lambda potential")
+    SECTION("lambda potential (i)")
     {
-        auto ff =
-            md::make_plane_inward_forcefield([](md::index i) {
-                return md::harmonic_potential{i * 1.0};
-            })
-            .set_plane(
-                md::plane{md::vector{0, 0, 1}, md::point{0, 0, 0}}
-            );
+        auto ff = md::make_plane_inward_forcefield(
+            [](md::index) {
+                return md::harmonic_potential{42};
+            }
+        );
 
         md::system system;
-        md::harmonic_potential pot1 = ff.plane_inward_potential(system, 1);
-        md::harmonic_potential pot2 = ff.plane_inward_potential(system, 2);
-        CHECK(pot1.spring_constant == Approx(1));
-        CHECK(pot2.spring_constant == Approx(2));
+        md::harmonic_potential pot = ff.plane_inward_potential(system, 0);
+        CHECK(pot.spring_constant == 42);
+    }
 
-        using ff_type = decltype(ff);
-        CHECK(std::is_base_of<md::plane_surface_forcefield<ff_type>, ff_type>::value);
+    SECTION("lambda potential (system, i)")
+    {
+        auto ff = md::make_plane_inward_forcefield(
+            [](md::system const&, md::index) {
+                return md::harmonic_potential{42};
+            }
+        );
+
+        md::system system;
+        md::harmonic_potential pot = ff.plane_inward_potential(system, 0);
+        CHECK(pot.spring_constant == 42);
     }
 
     SECTION("lambda plane")
     {
         auto ff =
-            md::make_plane_outward_forcefield([](md::index i) {
-                return md::harmonic_potential{i * 1.0};
-            })
+            md::make_plane_outward_forcefield(md::harmonic_potential{})
             .set_plane([&] {
                 return md::plane{md::vector{1, 2, 3}, md::point{4, 5, 6}};
             });
@@ -248,47 +249,48 @@ TEST_CASE("make_plane_outward_forcefield - creates a plane_surface_forcefield")
 {
     SECTION("fixed potential")
     {
-        md::harmonic_potential potential{1.23};
-        auto ff =
-            md::make_plane_outward_forcefield(potential)
-            .set_plane(
-                md::plane{md::vector{0, 0, 1}, md::point{0, 0, 0}}
-            );
+        auto ff = md::make_plane_outward_forcefield(
+            md::harmonic_potential{42}
+        );
 
         md::system system;
         md::harmonic_potential pot = ff.plane_outward_potential(system, 0);
-        CHECK(pot.spring_constant == Approx(potential.spring_constant));
+        CHECK(pot.spring_constant == 42);
 
         using ff_type = decltype(ff);
         CHECK(std::is_base_of<md::plane_surface_forcefield<ff_type>, ff_type>::value);
     }
 
-    SECTION("lambda potential")
+    SECTION("lambda potential (i)")
     {
-        auto ff =
-            md::make_plane_outward_forcefield([](md::index i) {
-                return md::harmonic_potential{i * 1.0};
-            })
-            .set_plane(
-                md::plane{md::vector{0, 0, 1}, md::point{0, 0, 0}}
-            );
+        auto ff = md::make_plane_outward_forcefield(
+            [](md::index) {
+                return md::harmonic_potential{42};
+            }
+        );
 
         md::system system;
-        md::harmonic_potential pot1 = ff.plane_outward_potential(system, 1);
-        md::harmonic_potential pot2 = ff.plane_outward_potential(system, 2);
-        CHECK(pot1.spring_constant == Approx(1));
-        CHECK(pot2.spring_constant == Approx(2));
+        md::harmonic_potential pot = ff.plane_outward_potential(system, 0);
+        CHECK(pot.spring_constant == 42);
+    }
 
-        using ff_type = decltype(ff);
-        CHECK(std::is_base_of<md::plane_surface_forcefield<ff_type>, ff_type>::value);
+    SECTION("lambda potential (system, i)")
+    {
+        auto ff = md::make_plane_outward_forcefield(
+            [](md::system const&, md::index) {
+                return md::harmonic_potential{42};
+            }
+        );
+
+        md::system system;
+        md::harmonic_potential pot = ff.plane_outward_potential(system, 0);
+        CHECK(pot.spring_constant == 42);
     }
 
     SECTION("lambda plane")
     {
         auto ff =
-            md::make_plane_outward_forcefield([](md::index i) {
-                return md::harmonic_potential{i * 1.0};
-            })
+            md::make_plane_outward_forcefield(md::harmonic_potential{})
             .set_plane([&] {
                 return md::plane{md::vector{1, 2, 3}, md::point{4, 5, 6}};
             });