+++ tests/forcefield/test_point_source_forcefield.cc
@@ -114,32 +114,41 @@ TEST_CASE("make_point_source_forcefield - creates a point_source_forcefield")
 {
     SECTION("fixed potential")
     {
-        auto ff = md::make_point_source_forcefield(md::harmonic_potential{1.23});
+        auto ff = md::make_point_source_forcefield(md::harmonic_potential{42});
         ff.set_point_source(md::point{1, 2, 3});
 
         md::system system;
         md::harmonic_potential pot = ff.point_source_potential(system, 0);
-        CHECK(pot.spring_constant == 1.23);
+        CHECK(pot.spring_constant == 42);
 
         using ff_type = decltype(ff);
         CHECK(std::is_base_of<md::point_source_forcefield<ff_type>, ff_type>::value);
     }
 
-    SECTION("lambda potential")
+    SECTION("lambda potential (i)")
     {
-        auto ff = md::make_point_source_forcefield([](md::index i) {
-            return md::harmonic_potential{i * 1.0};
-        });
-        ff.set_point_source(md::point{1, 2, 3});
+        auto ff = md::make_point_source_forcefield(
+            [](md::index) {
+                return md::harmonic_potential{42};
+            }
+        );
 
         md::system system;
-        md::harmonic_potential pot1 = ff.point_source_potential(system, 1);
-        md::harmonic_potential pot2 = ff.point_source_potential(system, 2);
-        CHECK(pot1.spring_constant == Approx(1));
-        CHECK(pot2.spring_constant == Approx(2));
+        md::harmonic_potential pot = ff.point_source_potential(system, 0);
+        CHECK(pot.spring_constant == 42);
+    }
 
-        using ff_type = decltype(ff);
-        CHECK(std::is_base_of<md::point_source_forcefield<ff_type>, ff_type>::value);
+    SECTION("lambda potential (system, i)")
+    {
+        auto ff = md::make_point_source_forcefield(
+            [](md::system const&, md::index) {
+                return md::harmonic_potential{42};
+            }
+        );
+
+        md::system system;
+        md::harmonic_potential pot = ff.point_source_potential(system, 0);
+        CHECK(pot.spring_constant == 42);
     }
 }