+++ CHANGELOG.md
@@ -11,6 +11,11 @@
   - Added `simulate_langevin_dynamics()`: A function that simulates Langevin
     dynamics of a system.
 - Forcefield templates:
+  - Added support for `(system, i)`-style lambdas as an argument for
+    `make_ellipsoid_inward_forcefield`, `make_ellipsoid_outward_forcefield`,
+    `make_sphere_inward_forcefield`, `make_sphere_outward_forcefield`,
+    `make_plane_inward_forcefield`, `make_plane_outward_forcefield` and
+    `make_point_source_forcefield`.
   - Added support for `(system, i, j)`-style lambdas as an argument for
     `make_neighbor_pairwise_forcefield`, `make_bruteforce_pairwise_forcefield`
     and `make_bonded_pairwise_forcefield`.